requirements.txt
asgiref==3.8.1
azure-core==1.32.0
azure-cosmos==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
Django==5.1.2
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.41
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
passlib==1.7.4
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
